requirements.txt
praw==7.8.1
PyYAML==6.0.2
retry==0.9.2
scipy==1.10.0
beautifulsoup4==4.9.1
Unidecode==1.1.1
lxml==4.9.1
requests==2.32.0